Refactoring, removing unused using statements Moved Colors from IUserBase.cs.cs to INotifyUI.cs
