+++ Majesty/Communication/SocketHandlerFactory.cs
@@ -1,8 +1,6 @@
 ﻿using Majesty.Communication.Sockets;
 using System;
-using System.Collections.Generic;
 using System.Net.Sockets;
-using System.Text;
 
 namespace Majesty.Communication
 {
