+++ Majesty/Communication/Sockets/SocketHandler.cs
@@ -1,6 +1,5 @@
 ﻿using Majesty.Messages;
 using System;
-using System.Collections.Generic;
 using System.Net.Sockets;
 using System.Security;
 using System.Text;
