+++ Majesty/Users/IUser.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Majesty.Users
+﻿namespace Majesty.Users
 {
     interface IUser<T> : IUserBase
     {
