+++ Majesty/Users/IUserBase.cs
@@ -1,23 +1,8 @@
 ﻿using Majesty.Messages;
-using System;
-using System.Collections.Generic;
-using System.Text;
+using Majesty.UI;
 
 namespace Majesty.Users
 {
-    public enum Colors
-    {
-        Hot_Pink = 0,
-        Lime = 1,
-        Goldenrod = 2,
-        Tomato = 3,
-        Medium_Orchid = 4,
-        Aquamarine = 5,
-        Alice_Blue = 6,
-        Dark_Salmon = 7,
-        Crimson = 8
-    }
-
     interface IUserBase
     {
         string NickName { get; }
